General purpose not working
I tried to create arrays of paths in order to have all of them. The for loop at line  126 should start with k=1, but in that case there are exceptions due to the connectivity problem.
Making it start with 2 should make everything work as before, but idk why this does not happen.